requirements.txt
Django==1.6.5
PyYAML==3.11
South==1.0